requirements.txt
# Core RAG dependencies
supabase==2.10.0
httpx==0.26.0
sentence-transformers==2.7.0
langchain==0.1.20
langchain-community==0.0.38
pgvector==0.2.5

# LLM providers
openai==1.30.1  # For OpenRouter API (OpenAI-compatible)

# Data collection
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11  # For RSS feeds
pyyaml==6.0.1  # For blog scraper configs
pytrends  # Google Trends API
huggingface_hub  # For HuggingFace model data with descriptions

# Data processing
# pandas==2.0.3
# numpy==1.24.3

# UI
streamlit==1.33.0
plotly==5.20.0  # For interactive charts

# Evaluation (Phase 4)
# ragas==0.1.7

# Utilities
python-dotenv==1.0.1
# pydantic==2.7.1
